Crate ic_validator

A crate for validating an HttpRequest.

An HttpRequest is considered valid iff:

The request hasn't expired relative to current_time.

The delegations haven't expired relative to current_time.

The signatures are corrrect.

Enums

AuthenticationError
Error in verifying the signature or authentication part of a request.

CanisterIdSet
An enum representing a set of canister IDs.

RequestValidationError
Error in validating an HttpRequest.

Functions

get_authorized_canisters
Returns the set of canisters that the request is authorized to act on.

validate_request
Validates the request and that the sender is authorized to send a message to the receiving canister.